feat: optimize commit-from-buffer instruction using Pinocchio for lower CU consumption (#97)
It is the fourth PR in the current stack. See #94 (the first PR in the stack) for a detailed description.

I couldn't test it, as it's not currently being invoked by the integration tests.

This PR rewrites `process_commit_state_from_buffer()`, which is very similar to `process_commit_state()` &mdash; in fact both of them are implemented in terms of `process_commit_state_internal()`. So its performance improvement should be similar to the performance improvement seen for `process_commit_state()` (see #96).